# backend/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateGroupDto": {
        "properties": {
          "name": {
            "description": "Human-readable roommate group name",
            "example": "Apartment 12A",
            "maxLength": 120,
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "type": "object"
      },
...
      "GroupMemberRole": {
        "description": "Target role for the member",
        "enum": [
          "ADMIN",
          "MEMBER"
        ],
        "type": "string"
      },
      "JoinGroupDto": {
        "properties": {
          "joinCode": {
            "description": "Join code generated by group admin",
            "example": "AB12CD34",
            "maxLength": 20,
            "minLength": 4,
            "type": "string"
          }
        },
        "required": [
          "joinCode"
        ],
        "type": "object"
      },
...
      "UpdateMemberRoleDto": {
        "properties": {
          "role": {
            "$ref": "#/components/schemas/GroupMemberRole"
          }
        },
        "required": [
          "role"
        ],
        "type": "object"
      }
    },
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "description": "Supabase access token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "Module 1-2 platform and authentication",
    "title": "RoomieManager Backend API",
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/groups": {
      "get": {
        "operationId": "GroupsController_listUserGroups",
        "parameters": [
          {
            "description": "1-based page number.",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "example": 1,
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "description": "Maximum number of items per page.",
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "example": 20,
              "maximum": 100,
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "description": "Optional sort field for current-user group listing.",
            "in": "query",
            "name": "sortBy",
            "required": false,
            "schema": {
              "enum": [
                "updatedAt",
                "createdAt",
                "name",
                "joinedAt"
              ],
              "type": "string"
            }
          },
          {
            "description": "Sort order applied to the selected field.",
            "in": "query",
            "name": "sortOrder",
            "required": false,
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "data": {
                      "groups": [
                        {
                          "createdAt": "2026-03-05T16:00:00.000Z",
                          "createdBy": "550e8400-e29b-41d4-a716-446655440001",
                          "id": "cm8z9ab120001mk8z4og1j0e9",
                          "joinCode": "ABCD1234",
                          "memberCount": 3,
                          "memberRole": "ADMIN",
                          "memberStatus": "ACTIVE",
                          "name": "Apartment 12A",
                          "updatedAt": "2026-03-05T16:05:00.000Z"
                        }
                      ],
                      "pagination": {
                        "hasNextPage": false,
                        "hasPreviousPage": false,
                        "page": 1,
                        "pageSize": 20,
                        "totalItems": 1,
                        "totalPages": 1
                      }
                    },
                    "meta": {
                      "requestId": "ad86d8f4-8f30-4383-9534-dbc56f5aa1af",
                      "timestamp": "2026-03-05T16:06:00.000Z"
                    },
                    "success": true
                  }
                }
              }
            },
            "description": "Returns active groups for the current user with pagination metadata."
          },
          "400": {
            "description": "Invalid pagination/sort query values."
          },
          "401": {
            "description": "Missing or invalid bearer token."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List groups for current active memberships.",
        "tags": [
          "Groups"
        ]
      },
      "post": {
        "operationId": "GroupsController_createGroup",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGroupDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Group created with admin membership and join code."
          },
          "401": {
            "description": "Missing or invalid bearer token."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a roommate group. Creator becomes admin.",
        "tags": [
          "Groups"
        ]
      }
    },
    "/api/v1/groups/{groupId}": {
      "get": {
        "operationId": "GroupsController_getGroup",
        "parameters": [
          {
            "in": "path",
            "name": "groupId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns group metadata and caller membership context."
          },
          "401": {
            "description": "Missing or invalid bearer token."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get group summary for current member.",
        "tags": [
          "Groups"
        ]
      }
    },
...
    "/api/v1/groups/{groupId}/dashboard": {
      "get": {
        "operationId": "GroupsController_getGroupDashboard",
        "parameters": [
          {
            "in": "path",
            "name": "groupId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "data": {
                      "chores": {
                        "assignedToMePendingCount": 2,
                        "completedCount": 8,
                        "overdueCount": 1,
                        "pendingCount": 4
                      },
                      "contract": {
                        "hasDraft": true,
                        "publishedVersion": 2,
                        "updatedAt": "2026-03-05T11:00:00.000Z"
                      },
                      "finance": {
                        "billCount": 5,
                        "latestBillIncurredAt": "2026-03-05T12:00:00.000Z",
                        "latestPaymentPaidAt": "2026-03-05T13:00:00.000Z",
                        "paymentCount": 7
                      },
                      "group": {
                        "createdAt": "2026-03-05T16:00:00.000Z",
                        "createdBy": "550e8400-e29b-41d4-a716-446655440001",
                        "id": "cm8z9ab120001mk8z4og1j0e9",
                        "joinCode": "ABCD1234",
                        "memberCount": 3,
                        "memberRole": "ADMIN",
                        "memberStatus": "ACTIVE",
                        "name": "Apartment 12A",
                        "updatedAt": "2026-03-05T16:05:00.000Z"
                      },
                      "members": {
                        "adminCount": 1,
                        "memberCount": 2,
                        "totalActive": 3
                      }
                    },
                    "meta": {
                      "requestId": "ad86d8f4-8f30-4383-9534-dbc56f5aa1af",
                      "timestamp": "2026-03-05T16:06:00.000Z"
                    },
                    "success": true
                  }
                }
              }
            },
            "description": "Returns aggregated group dashboard data for active members."
          },
          "400": {
            "description": "Invalid group ID format."
          },
          "401": {
            "description": "Missing or invalid bearer token."
          },
          "403": {
            "description": "Caller is not an active member of this group."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get frontend dashboard aggregates for a group (member/admin counts, chore counters, finance counters, contract status).",
        "tags": [
          "Groups"
        ]
      }
    },
    "/api/v1/groups/{groupId}/join-code/reset": {
      "post": {
        "operationId": "GroupsController_resetJoinCode",
        "parameters": [
          {
            "in": "path",
            "name": "groupId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns newly generated join code."
          },
          "401": {
            "description": "Missing or invalid bearer token."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Reset group join code (admin only).",
        "tags": [
          "Groups"
        ]
      }
    },
    "/api/v1/groups/{groupId}/members": {
      "get": {
        "operationId": "GroupsController_getGroupMembers",
        "parameters": [
          {
            "in": "path",
            "name": "groupId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "1-based page number.",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "example": 1,
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "description": "Maximum number of items per page.",
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "example": 20,
              "maximum": 100,
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "description": "Optional sort field for group members.",
            "in": "query",
            "name": "sortBy",
            "required": false,
            "schema": {
              "enum": [
                "role",
                "joinedAt",
                "createdAt"
              ],
              "type": "string"
            }
          },
          {
            "description": "Sort order applied to the selected field.",
            "in": "query",
            "name": "sortOrder",
            "required": false,
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "data": {
                      "groupId": "cm8w9z0abc123def456ghi789",
                      "members": [
                        {
                          "createdAt": "2026-03-05T14:40:00.000Z",
                          "joinedAt": "2026-03-05T14:40:00.000Z",
                          "role": "ADMIN",
                          "status": "ACTIVE",
                          "updatedAt": "2026-03-05T14:40:00.000Z",
                          "userId": "550e8400-e29b-41d4-a716-446655440001"
                        }
                      ],
                      "pagination": {
                        "hasNextPage": false,
                        "hasPreviousPage": false,
                        "page": 1,
                        "pageSize": 20,
                        "totalItems": 1,
                        "totalPages": 1
                      }
                    },
                    "meta": {
                      "requestId": "f073d9ed-7d9f-4902-a8ee-bfca8de3e4ff",
                      "timestamp": "2026-03-05T14:41:00.000Z"
                    },
                    "success": true
                  }
                }
              }
            },
            "description": "Returns current group members with roles, statuses, and pagination metadata."
          },
          "400": {
            "description": "Invalid group ID or invalid pagination/sort query values."
          },
          "401": {
            "description": "Missing or invalid bearer token."
          },
          "403": {
            "description": "Caller is not an active member of this group."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List active members in the group.",
        "tags": [
          "Groups"
        ]
      }
    },
    "/api/v1/groups/{groupId}/members/{userId}": {
      "delete": {
        "operationId": "GroupsController_removeMember",
        "parameters": [
          {
            "in": "path",
            "name": "groupId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Marks target member as inactive and returns result."
          },
          "401": {
            "description": "Missing or invalid bearer token."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove member from group (admin only).",
        "tags": [
          "Groups"
        ]
      }
    },
    "/api/v1/groups/{groupId}/members/{userId}/role": {
      "patch": {
        "operationId": "GroupsController_updateMemberRole",
        "parameters": [
          {
            "in": "path",
            "name": "groupId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMemberRoleDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Returns updated member role state."
          },
          "401": {
            "description": "Missing or invalid bearer token."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update group member role (admin only).",
        "tags": [
          "Groups"
        ]
      }
    },
...
    "/api/v1/groups/join": {
      "post": {
        "operationId": "GroupsController_joinGroup",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinGroupDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Joins group as member and returns group context."
          },
          "401": {
            "description": "Missing or invalid bearer token."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Join a group using join code.",
        "tags": [
          "Groups"
        ]
      }
    },
...
  },
  "servers": [],
...
}